remove interactive command
