+++ go/cmd/flags/flags.go
@@ -22,4 +22,5 @@ const (
 	RCIncrement  = "rc"
 	RunLive      = "live"
 	VtOpRelease  = "vtop-release"
+	Help         = "help"
 )
